requirements.txt
astroid==2.3.3
attrs==19.3.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
google-api-core==1.31.1
google-auth==1.34.0
google-cloud-bigquery==1.24.0
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
google-auth-oauthlib==0.4.4
google-auth-httplib2==0.1.0
google-api-python-client==2.0.2
idna==2.9
importlib-metadata==1.5.0
isort==4.3.21
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.4.4
pyparsing==2.4.6
pytest==5.4.1
pytz==2019.3
PyYAML==5.4.1
requests==2.23.0
rope==0.16.0
rsa==4.0
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.8
wcwidth==0.1.8
wrapt==1.11.2
pyOpenSSL==21.0.0
protobuf==3.12.0
cerberus==1.3.4
pandas==1.0
numpy==1.21.1
simplejson==3.16.0
pandas-gbq==0.12.0
gcsfs==0.6.0
s3fs==0.4.0
pyarrow==4.0
xlrd==1.2.0
python-dateutil==2.8.2
httplib2==0.19.1
testresources==2.0.1
pyminizip==0.2.4
lxml==4.7.1
paramiko==2.6.0
ndjson==0.3.1
